requirements.txt
Django==5.1.4
channels==4.2.0
channels_redis==4.2.1
redis==5.2.1
uvicorn==0.34.0
uvloop==0.21.0
websockets==14.1
wsproto==1.2.0
djangorestframework==3.15.2